@@ -5,19 +5,26 @@ import (
 	"encoding/json"
 	"fmt"
 	config "github.com/OliveTin/OliveTin/internal/config"
+	"github.com/OliveTin/OliveTin/internal/filehelper"
 	sv "github.com/OliveTin/OliveTin/internal/stringvariables"
-	"github.com/fsnotify/fsnotify"
 	log "github.com/sirupsen/logrus"
-	"github.com/spf13/viper"
 	"gopkg.in/yaml.v3"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 )
 
+var (
+	EntityChangedSender chan bool
+	listeners           []func()
+)
+
+func AddListener(l func()) {
+	listeners = append(listeners, l)
+}
+
 func SetupEntityFileWatchers(cfg *config.Config) {
-	configDir := filepath.Dir(viper.ConfigFileUsed())
+	configDir := cfg.GetDir()
 
 	configDirVar := filepath.Join(configDir, "var") // for development purposes
 
@@ -36,71 +43,11 @@ func SetupEntityFileWatchers(cfg *config.Config) {
 			}).Debugf("Adding config dir to entity file path")
 		}
 
-		go watch(p, ef.Name)
+		go filehelper.WatchFileWrite(p, func(filename string) {
+			loadEntityFile(p, ef.Name)
+		})
 
 		loadEntityFile(p, ef.Name)
-	}
-}
-
-func watch(file string, entityname string) {
-	log.WithFields(log.Fields{
-		"file": file,
-		"name": entityname,
-	}).Infof("Watching entity file")
-
-	watcher, err := fsnotify.NewWatcher()
-
-	if err != nil {
-		log.Errorf("Could not watch entity file: %v", err)
-		return
-	}
-
-	defer watcher.Close()
-
-	done := make(chan bool)
-
-	go func() {
-		for {
-			processEvent(watcher, file, entityname)
-		}
-	}()
-
-	err = watcher.Add(file)
-
-	if err != nil {
-		log.WithFields(log.Fields{
-			"file": file,
-		}).Errorf("Could not create entity watcher: %v", err)
-	}
-
-	<-done
-}
-
-func processEvent(watcher *fsnotify.Watcher, filename string, entityname string) {
-	select {
-	case event, ok := <-watcher.Events:
-		if !ok {
-			return
-		}
-
-		loadEntityFileIfWritten(&event, filename, entityname)
-
-		return
-	case err := <-watcher.Errors:
-		log.Errorf("Error in fsnotify: %v", err)
-		return
-	}
-}
-
-func loadEntityFileIfWritten(event *fsnotify.Event, filename string, entityname string) {
-	if event.Has(fsnotify.Remove) {
-		log.WithFields(log.Fields{
-			"file": filename,
-		}).Warnf("Entity file deleted! Will no longer be able to watch for changes!")
-	}
-
-	if event.Has(fsnotify.Write) {
-		loadEntityFile(filename, entityname)
 	}
 }
 
@@ -118,7 +65,7 @@ func loadEntityFileJson(filename string, entityname string) {
 		"name": entityname,
 	}).Infof("Loading entity file with JSON format")
 
-	jfile, err := ioutil.ReadFile(filename)
+	jfile, err := os.ReadFile(filename)
 
 	if err != nil {
 		log.Errorf("ReadIn: %v", err)
@@ -142,7 +89,7 @@ func loadEntityFileJson(filename string, entityname string) {
 		data = append(data, d)
 	}
 
-	updateEvmFromFile(entityname, data)
+	updateSvFromFile(entityname, data)
 }
 
 func loadEntityFileYaml(filename string, entityname string) {
@@ -151,7 +98,7 @@ func loadEntityFileYaml(filename string, entityname string) {
 		"name": entityname,
 	}).Infof("Loading entity file with YAML format")
 
-	yfile, err := ioutil.ReadFile(filename)
+	yfile, err := os.ReadFile(filename)
 
 	if err != nil {
 		log.Errorf("ReadIn: %v", err)
@@ -166,21 +113,27 @@ func loadEntityFileYaml(filename string, entityname string) {
 		log.Errorf("Unmarshal: %v", err)
 	}
 
-	updateEvmFromFile(entityname, data)
+	updateSvFromFile(entityname, data)
 }
 
-func updateEvmFromFile(entityname string, data []map[string]string) {
+func updateSvFromFile(entityname string, data []map[string]string) {
+	log.Debugf("updateSvFromFile: %+v", data)
+
 	count := len(data)
 
 	sv.RemoveKeysThatStartWith("entities." + entityname)
 
-	sv.Contents["entities."+entityname+".count"] = fmt.Sprintf("%v", count)
+	sv.SetEntityCount(entityname, count)
 
 	for i, mapp := range data {
 		prefix := "entities." + entityname + "." + fmt.Sprintf("%v", i)
 
 		for k, v := range mapp {
-			sv.Contents[prefix+"."+k] = v
+			sv.Set(prefix+"."+k, v)
 		}
+	}
+
+	for _, l := range listeners {
+		l()
 	}
 }
