@@ -17,6 +17,18 @@ import (
 	"net/url"
 )
 
+func logDebugRequest(cfg *config.Config, source string, r *http.Request) {
+	if cfg.LogDebugOptions.SingleFrontendRequests {
+		log.Debugf("SingleFrontend HTTP Req URL %v: %q", source, r.URL)
+
+		if cfg.LogDebugOptions.SingleFrontendRequestHeaders {
+			for name, values := range r.Header {
+				log.Debugf("SingleFrontend HTTP Req Hdr: %v = %v", name, values)
+			}
+		}
+	}
+}
+
 // StartSingleHTTPFrontend will create a reverse proxy that proxies the API
 // and webui internally.
 func StartSingleHTTPFrontend(cfg *config.Config) {
@@ -33,19 +45,33 @@ func StartSingleHTTPFrontend(cfg *config.Config) {
 	mux := http.NewServeMux()
 
 	mux.HandleFunc("/api/", func(w http.ResponseWriter, r *http.Request) {
-		log.Debugf("api req: %q", r.URL)
+		logDebugRequest(cfg, "api ", r)
+
 		apiProxy.ServeHTTP(w, r)
 	})
 
 	mux.HandleFunc("/websocket", func(w http.ResponseWriter, r *http.Request) {
-		log.Debugf("ws  req: %q", r.URL)
+		logDebugRequest(cfg, "ws  ", r)
+
 		websocket.HandleWebsocket(w, r)
 	})
 
 	mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		log.Debugf("ui  req: %q", r.URL)
+		logDebugRequest(cfg, "ui  ", r)
+
 		webuiProxy.ServeHTTP(w, r)
 	})
+
+	if cfg.Prometheus.Enabled {
+		promURL, _ := url.Parse("http://" + cfg.ListenAddressPrometheus)
+		promProxy := httputil.NewSingleHostReverseProxy(promURL)
+
+		mux.HandleFunc("/metrics", func(w http.ResponseWriter, r *http.Request) {
+			logDebugRequest(cfg, "prom", r)
+
+			promProxy.ServeHTTP(w, r)
+		})
+	}
 
 	srv := &http.Server{
 		Addr:    cfg.ListenAddressSingleHTTPFrontend,
