@@ -6,11 +6,11 @@ import (
 )
 
 func init() {
-	sv.Contents["OliveTin.build.commit"] = Build.Commit
-	sv.Contents["OliveTin.build.version"] = Build.Version
-	sv.Contents["OliveTin.build.date"] = Build.Date
-	sv.Contents["OliveTin.runtime.os"] = Runtime.OS
-	sv.Contents["OliveTin.runtime.os.pretty"] = Runtime.OSReleasePrettyName
-	sv.Contents["OliveTin.runtime.arch"] = Runtime.Arch
-	sv.Contents["OliveTin.runtime.incontainer"] = fmt.Sprintf("%v", Runtime.InContainer)
+	sv.Set("OliveTin.build.commit", Build.Commit)
+	sv.Set("OliveTin.build.version", Build.Version)
+	sv.Set("OliveTin.build.date", Build.Date)
+	sv.Set("OliveTin.runtime.os", Runtime.OS)
+	sv.Set("OliveTin.runtime.os.pretty", Runtime.OSReleasePrettyName)
+	sv.Set("OliveTin.runtime.arch", Runtime.Arch)
+	sv.Set("OliveTin.runtime.incontainer", fmt.Sprintf("%v", Runtime.InContainer))
 }
