@@ -1,86 +1,57 @@
 package onfileindir
 
 import (
+	"fmt"
 	"github.com/OliveTin/OliveTin/internal/acl"
 	"github.com/OliveTin/OliveTin/internal/config"
 	"github.com/OliveTin/OliveTin/internal/executor"
-	"github.com/fsnotify/fsnotify"
-	log "github.com/sirupsen/logrus"
+	"github.com/OliveTin/OliveTin/internal/filehelper"
+	"os"
+	"path/filepath"
 )
 
 func WatchFilesInDirectory(cfg *config.Config, ex *executor.Executor) {
 	for _, action := range cfg.Actions {
 		for _, dirname := range action.ExecOnFileChangedInDir {
-			watch(dirname, action, cfg, ex, fsnotify.Write)
+			go filehelper.WatchDirectoryWrite(dirname, func(filename string) {
+				scheduleExec(action, cfg, ex, filename)
+			})
 		}
 
 		for _, dirname := range action.ExecOnFileCreatedInDir {
-			watch(dirname, action, cfg, ex, fsnotify.Create)
+			go filehelper.WatchDirectoryCreate(dirname, func(filename string) {
+				scheduleExec(action, cfg, ex, filename)
+			})
 		}
 	}
 }
 
-func watch(directory string, action *config.Action, cfg *config.Config, ex *executor.Executor, eventType fsnotify.Op) {
-	log.WithFields(log.Fields{
-		"dir":       directory,
-		"eventType": eventType,
-	}).Infof("Watching dir")
-
-	watcher, err := fsnotify.NewWatcher()
-
-	if err != nil {
-		log.Errorf("Could not watch for files being created: %v", err)
-		return
+func scheduleExec(action *config.Action, cfg *config.Config, ex *executor.Executor, path string) {
+	args := map[string]string{
+		"filepath": path,
+		"filename": filepath.Base(path),
+		"filedir":  filepath.Dir(path),
+		"fileext":  filepath.Ext(path),
 	}
 
-	defer watcher.Close()
-
-	done := make(chan bool)
-
-	go func() {
-		for {
-			processEvent(watcher, action, cfg, ex, eventType)
-		}
-	}()
-
-	err = watcher.Add("/tmp")
-
-	if err != nil {
-		log.Errorf("Could not create watcher: %v", err)
+	if stat, err := os.Stat(path); err == nil {
+		args["filesizebytes"] = fmt.Sprintf("%v", stat.Size())
+		args["filemode"] = fmt.Sprintf("%#o", stat.Mode())
+		args["filemtime"] = fmt.Sprintf("%v", stat.ModTime())
+		args["fileisdir"] = fmt.Sprintf("%v", stat.IsDir())
 	}
 
-	<-done
-}
+	fmt.Printf("%+v", args)
 
-func processEvent(watcher *fsnotify.Watcher, action *config.Action, cfg *config.Config, ex *executor.Executor, eventType fsnotify.Op) {
-	select {
-	case event, ok := <-watcher.Events:
-		if !ok {
-			return
-		}
-
-		checkEvent(&event, action, cfg, ex, eventType)
-		break
-	case err := <-watcher.Errors:
-		log.Errorf("Error in fsnotify: %v", err)
-		return
+	req := &executor.ExecutionRequest{
+		ActionTitle: action.Title,
+		Cfg:         cfg,
+		Tags:        []string{"fileindir"},
+		Arguments:   args,
+		AuthenticatedUser: &acl.AuthenticatedUser{
+			Username: "fileindir",
+		},
 	}
-}
 
-func checkEvent(event *fsnotify.Event, action *config.Action, cfg *config.Config, ex *executor.Executor, eventType fsnotify.Op) {
-	if event.Has(eventType) {
-		req := &executor.ExecutionRequest{
-			ActionTitle: action.Title,
-			Cfg:         cfg,
-			Tags:        []string{"fileindir"},
-			Arguments: map[string]string{
-				"filename": event.Name,
-			},
-			AuthenticatedUser: &acl.AuthenticatedUser{
-				Username: "fileindir",
-			},
-		}
-
-		ex.ExecRequest(req)
-	}
+	ex.ExecRequest(req)
 }
