@@ -9,18 +9,68 @@
 
 package stringvariables
 
-var Contents map[string]string
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"strings"
+	"sync"
+)
+
+var (
+	contents map[string]string
+
+	metricSvCount = promauto.NewGauge(prometheus.GaugeOpts{
+		Name: "olivetin_sv_count",
+		Help: "The number entries in the sv map",
+	})
+
+	rwmutex = sync.RWMutex{}
+)
 
 func init() {
-	Contents = make(map[string]string)
+	rwmutex.Lock()
+
+	contents = make(map[string]string)
+
+	rwmutex.Unlock()
 }
 
 func Get(key string) string {
-	v, ok := Contents[key]
+	rwmutex.RLock()
+
+	v, ok := contents[key]
+
+	rwmutex.RUnlock()
 
 	if !ok {
 		return ""
 	} else {
 		return v
 	}
+}
+
+func GetAll() map[string]string {
+	return contents
+}
+
+func Set(key string, value string) {
+	rwmutex.Lock()
+
+	contents[key] = value
+
+	metricSvCount.Set(float64(len(contents)))
+
+	rwmutex.Unlock()
+}
+
+func RemoveKeysThatStartWith(search string) {
+	rwmutex.Lock()
+
+	for k, _ := range contents {
+		if strings.HasPrefix(k, search) {
+			delete(contents, k)
+		}
+	}
+
+	rwmutex.Unlock()
 }
