@@ -8,11 +8,16 @@ import (
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/reflect/protoreflect"
 	"net/http"
+	"sync"
 )
 
 var upgrader = ws.Upgrader{
 	CheckOrigin: checkOriginPermissive,
 }
+
+var (
+	sendmutex = sync.Mutex{}
+)
 
 type WebsocketClient struct {
 	conn *ws.Conn
@@ -36,6 +41,14 @@ func (WebsocketExecutionListener) OnExecutionStarted(title string) {
 			Action: title,
 		});
 	*/
+}
+
+func OnEntityChanged() {
+	broadcast(&pb.EventEntityChanged{})
+}
+
+func (WebsocketExecutionListener) OnActionMapRebuilt() {
+	broadcast(&pb.EventConfigChanged{})
 }
 
 /*
@@ -67,28 +80,37 @@ func (WebsocketExecutionListener) OnOutputChunk(chunk []byte, executionTrackingI
 }
 
 func (WebsocketExecutionListener) OnExecutionFinished(logEntry *executor.InternalLogEntry) {
-	le := &pb.LogEntry{
-		ActionTitle:         logEntry.ActionTitle,
-		ActionIcon:          logEntry.ActionIcon,
-		ActionId:            logEntry.ActionId,
-		DatetimeStarted:     logEntry.DatetimeStarted,
-		DatetimeFinished:    logEntry.DatetimeFinished,
-		Stdout:              logEntry.Stdout,
-		Stderr:              logEntry.Stderr,
-		TimedOut:            logEntry.TimedOut,
-		Blocked:             logEntry.Blocked,
-		ExitCode:            logEntry.ExitCode,
-		Tags:                logEntry.Tags,
-		ExecutionTrackingId: logEntry.ExecutionTrackingID,
-		ExecutionStarted:    logEntry.ExecutionStarted,
-		ExecutionFinished:   logEntry.ExecutionFinished,
+	evt := &pb.EventExecutionFinished{
+		LogEntry: &pb.LogEntry{
+			ActionTitle:         logEntry.ActionTitle,
+			ActionIcon:          logEntry.ActionIcon,
+			ActionId:            logEntry.ActionId,
+			DatetimeStarted:     logEntry.DatetimeStarted.Format("2006-01-02 15:04:05"),
+			DatetimeFinished:    logEntry.DatetimeFinished.Format("2006-01-02 15:04:05"),
+			Stdout:              logEntry.Stdout,
+			Stderr:              logEntry.Stderr,
+			TimedOut:            logEntry.TimedOut,
+			Blocked:             logEntry.Blocked,
+			ExitCode:            logEntry.ExitCode,
+			Tags:                logEntry.Tags,
+			ExecutionTrackingId: logEntry.ExecutionTrackingID,
+			ExecutionStarted:    logEntry.ExecutionStarted,
+			ExecutionFinished:   logEntry.ExecutionFinished,
+		},
 	}
 
-	broadcast("ExecutionFinished", le)
+	broadcast(evt)
 }
 
-func broadcast(messageType string, pbmsg protoreflect.ProtoMessage) {
+func broadcast(pbmsg protoreflect.ProtoMessage) {
 	payload, err := marshalOptions.Marshal(pbmsg)
+
+	if err != nil {
+		log.Errorf("websocket marshal error: %v", err)
+		return
+	}
+
+	messageType := pbmsg.ProtoReflect().Descriptor().FullName()
 
 	// <EVIL>
 	// So, the websocket wants to encode messages using the same protomarshaller
@@ -109,14 +131,11 @@ func broadcast(messageType string, pbmsg protoreflect.ProtoMessage) {
 	hackyMessage = append(hackyMessage, []byte("}")...)
 	// </EVIL>
 
-	if err != nil {
-		log.Errorf("websocket marshal error: %v", err)
-		return
-	}
-
+	sendmutex.Lock()
 	for _, client := range clients {
 		client.conn.WriteMessage(ws.TextMessage, hackyMessage)
 	}
+	sendmutex.Unlock()
 }
 
 func (c *WebsocketClient) messageLoop() {
@@ -146,7 +165,11 @@ func HandleWebsocket(w http.ResponseWriter, r *http.Request) bool {
 		conn: c,
 	}
 
+	sendmutex.Lock()
+
 	clients = append(clients, wsclient)
+
+	sendmutex.Unlock()
 
 	go wsclient.messageLoop()
 
