@@ -12,13 +12,23 @@ export function initMarshaller () {
 
   window.logEntries = {}
 
-  window.addEventListener('ExecutionFinished', onExecutionFinished)
+  window.initialHash = window.location.hash
+
+  window.currentSection = ''
+
+  window.addEventListener('EventExecutionFinished', onExecutionFinished)
   window.addEventListener('OutputChunk', onOutputChunk)
 }
 
 export function marshalDashboardComponentsJsonToHtml (json) {
   marshalActionsJsonToHtml(json)
   marshalDashboardStructureToHtml(json)
+
+  document.getElementById('username').innerText = json.authenticatedUser
+
+  changeDirectory(null)
+
+  document.body.setAttribute('initial-marshal-complete', 'true')
 }
 
 function marshalActionsJsonToHtml (json) {
@@ -58,7 +68,7 @@ function onOutputChunk (evt) {
 }
 
 function onExecutionFinished (evt) {
-  const logEntry = evt.payload
+  const logEntry = evt.payload.logEntry
 
   const actionButton = window.actionButtons[logEntry.actionTitle]
 
@@ -99,6 +109,10 @@ function onExecutionFinished (evt) {
 }
 
 function showSection (title) {
+  title = title.replaceAll(' ', '')
+
+  window.currentSection = title
+
   for (const section of document.querySelectorAll('section')) {
     if (section.title === title) {
       section.style.display = 'block'
@@ -107,33 +121,59 @@ function showSection (title) {
     }
   }
 
-  for (const otherName of ['Actions', 'Logs']) {
-    document.getElementById('show' + otherName).classList.remove('activeSection')
-    document.getElementById('content' + otherName).hidden = true
-  }
-
-  //  document.getElementById('show' + name).classList.add('activeSection')
-  // document.getElementById('content' + name).hidden = false
-
-  document.getElementById('hide-sidebar-checkbox').checked = true
+  setSectionNavigationVisible(false)
 
   changeDirectory(null)
 }
 
+function setSectionNavigationVisible (visible) {
+  const nav = document.querySelector('nav')
+  const btn = document.getElementById('sidebar-toggler-button')
+
+  if (document.body.classList.contains('has-sidebar')) {
+    if (visible) {
+      btn.setAttribute('aria-pressed', false)
+      btn.setAttribute('aria-label', 'Open sidebar navigation')
+      btn.innerHTML = '&laquo;'
+
+      nav.classList.add('shown')
+      nav.style.display = 'flex'
+    } else {
+      btn.setAttribute('aria-pressed', true)
+      btn.setAttribute('aria-label', 'Close sidebar navigation')
+      btn.innerHTML = '&#9776;'
+
+      nav.classList.remove('shown')
+      setTimeout(() => {
+        nav.style.display = 'none'
+      }, 600)
+    }
+  } else {
+    btn.disabled = true
+  }
+}
+
 export function setupSectionNavigation (style) {
   const nav = document.querySelector('nav')
+  const btn = document.getElementById('sidebar-toggler-button')
 
   if (style === 'sidebar') {
-    nav.classList += 'sidebar'
+    nav.classList.add('sidebar')
 
-    document.body.classList += 'has-sidebar'
+    document.body.classList.add('has-sidebar')
+
+    btn.onclick = () => {
+      if (nav.classList.contains('shown')) {
+        setSectionNavigationVisible(false)
+      } else {
+        setSectionNavigationVisible(true)
+      }
+    }
   } else {
-    nav.classList += 'topbar'
+    nav.classList.add('topbar')
 
-    document.body.classList += 'has-topbar'
+    document.body.classList.add('has-topbar')
   }
-
-  nav.hidden = false
 
   document.getElementById('showActions').onclick = () => { showSection('Actions') }
   document.getElementById('showLogs').onclick = () => { showSection('Logs') }
@@ -143,14 +183,14 @@ function marshalDashboardStructureToHtml (json) {
   const nav = document.getElementById('navigation-links')
 
   for (const dashboard of json.dashboards) {
-    const oldsection = document.querySelector('section[title="' + dashboard.title + '"]')
+    const oldsection = document.querySelector('section[title="' + dashboard.title.replace(' ', '') + '"]')
 
     if (oldsection != null) {
       oldsection.remove()
     }
 
     const section = document.createElement('section')
-    section.title = dashboard.title
+    section.title = dashboard.title.replaceAll(' ', '')
 
     const def = createFieldset('default', section)
     section.appendChild(def)
@@ -167,8 +207,9 @@ function marshalDashboardStructureToHtml (json) {
     const navigationA = document.createElement('a')
     navigationA.title = dashboard.title
     navigationA.innerText = dashboard.title
+    navigationA.setAttribute('href', '#' + dashboard.title.replace(' ', ''))
     navigationA.onclick = () => {
-      showSection(dashboard.title)
+      showSection(dashboard.title.replace(' ', ''))
     }
 
     const navigationLi = document.createElement('li')
@@ -186,12 +227,18 @@ function marshalDashboardStructureToHtml (json) {
     }
   }
 
-  if (rootGroup.querySelectorAll('action-button').length === 0 && json.dashboards.length > 0) {
-    nav.querySelector('li[title="Actions"]').style.display = 'none'
-
-    showSection(json.dashboards[0].title)
+  if (window.currentSection !== '') {
+    showSection(window.currentSection)
+  } else if (window.initialHash !== '' && document.body.getAttribute('initial-marshal-complete') === null) {
+    showSection(window.initialHash.replace('#', ''))
   } else {
-    showSection('Actions')
+    if (rootGroup.querySelectorAll('action-button').length === 0 && json.dashboards.length > 0) {
+      nav.querySelector('li[title="Actions"]').style.display = 'none'
+
+      showSection(json.dashboards[0].title)
+    } else {
+      showSection('Actions')
+    }
   }
 }
 
@@ -362,12 +409,8 @@ function changeDirectory (selected) {
   document.title = title.innerText
 
   if (selected === null) {
-    window.location.hash = null
-
     window.history.pushState({ dir: null }, null, '#')
   } else {
-    window.location.hash = selected
-
     window.history.pushState({ dir: selected }, null, '#' + selected)
   }
 }
@@ -397,6 +440,7 @@ function createDirectoryBreadcrumb (title, link) {
 function marshalDisplay (item, fieldset) {
   const display = document.createElement('div')
   display.innerHTML = item.title
+  display.classList.add('display')
 
   fieldset.appendChild(display)
 }
@@ -494,3 +538,19 @@ window.addEventListener('popstate', (e) => {
     changeDirectory(e.state.dir)
   }
 })
+
+export function refreshServerConnectionLabel () {
+  if (window.restAvailable) {
+    document.querySelector('#serverConnectionRest').classList.remove('error')
+  } else {
+    document.querySelector('#serverConnectionRest').classList.add('error')
+  }
+
+  if (window.websocketAvailable) {
+    document.querySelector('#serverConnectionWebSocket').classList.remove('error')
+    document.querySelector('#serverConnectionWebSocket').innerText = 'WebSocket'
+  } else {
+    document.querySelector('#serverConnectionWebSocket').classList.add('error')
+    document.querySelector('#serverConnectionWebSocket').innerText = 'WebSocket Error'
+  }
+}
