@@ -1,6 +1,12 @@
 'use strict'
 
-import { initMarshaller, setupSectionNavigation, marshalDashboardComponentsJsonToHtml, marshalLogsJsonToHtml } from './js/marshaller.js'
+import {
+  initMarshaller,
+  setupSectionNavigation,
+  marshalDashboardComponentsJsonToHtml,
+  marshalLogsJsonToHtml,
+  refreshServerConnectionLabel
+} from './js/marshaller.js'
 import { checkWebsocketConnection } from './js/websocket.js'
 
 function searchLogs (e) {
@@ -47,20 +53,6 @@ function refreshLoop () {
   refreshServerConnectionLabel()
 }
 
-function refreshServerConnectionLabel () {
-  if (window.restAvailable) {
-    document.querySelector('#serverConnectionRest').classList.remove('error')
-  } else {
-    document.querySelector('#serverConnectionRest').classList.add('error')
-  }
-
-  if (window.websocketAvailable) {
-    document.querySelector('#serverConnectionWebSocket').classList.remove('error')
-  } else {
-    document.querySelector('#serverConnectionWebSocket').classList.add('error')
-  }
-}
-
 function fetchGetDashboardComponents () {
   window.fetch(window.restBaseUrl + 'GetDashboardComponents', {
     cors: 'cors'
@@ -68,7 +60,7 @@ function fetchGetDashboardComponents () {
     return res.json()
   }).then(res => {
     if (!window.restAvailable) {
-      window.clearBigErrors('fetch-buttons')
+      window.clearBigErrors()
     }
 
     window.restAvailable = true
@@ -77,7 +69,7 @@ function fetchGetDashboardComponents () {
     refreshServerConnectionLabel() // in-case it changed, update the label quicker
   }).catch((err) => { // err is 1st arg
     window.restAvailable = false
-    window.showBigError('fetch-buttons', 'getting buttons', err, 'blat')
+    window.showBigError('fetch-buttons', 'getting buttons', err, false)
   })
 }
 
@@ -89,21 +81,14 @@ function fetchGetLogs () {
   }).then(res => {
     marshalLogsJsonToHtml(res)
   }).catch(err => {
-    window.showBigError('fetch-buttons', 'getting buttons', err, 'blat')
+    window.showBigError('fetch-buttons', 'getting buttons', err, false)
   })
 }
 
 function processWebuiSettingsJson (settings) {
+  setupSectionNavigation(settings.SectionNavigationStyle)
+
   window.restBaseUrl = settings.Rest
-
-  if (settings.ThemeName) {
-    const themeCss = document.createElement('link')
-    themeCss.setAttribute('rel', 'stylesheet')
-    themeCss.setAttribute('type', 'text/css')
-    themeCss.setAttribute('href', '/themes/' + settings.ThemeName + '/theme.css')
-
-    document.head.appendChild(themeCss)
-  }
 
   document.querySelector('#currentVersion').innerText = settings.CurrentVersion
 
@@ -112,8 +97,13 @@ function processWebuiSettingsJson (settings) {
     document.querySelector('#available-version').hidden = false
   }
 
-  document.querySelector('#perma-widget').hidden = !settings.ShowNavigation
-  document.querySelector('footer[title="footer"]').hidden = !settings.ShowFooter
+  if (!settings.ShowNavigation) {
+    document.querySelector('header').style.display = 'none'
+  }
+
+  if (!settings.ShowFooter) {
+    document.querySelector('footer[title="footer"]').style.display = 'none'
+  }
 
   window.pageTitle = 'OliveTin'
 
@@ -130,7 +120,9 @@ function processWebuiSettingsJson (settings) {
 function main () {
   initMarshaller()
   setupLogSearchBox()
-  setupSectionNavigation('sidebar')
+
+  window.addEventListener('EventConfigChanged', fetchGetDashboardComponents)
+  window.addEventListener('EventEntityChanged', fetchGetDashboardComponents)
 
   window.fetch('webUiSettings.json').then(res => {
     return res.json()
